// client/src/app/shared/models/feedback.model.ts
export class Feedback {
  job_title: string;
  job_description: string;
  quoted_job_description: string;
  service_id: number;
  service_name: string;
  quality_of_work: number;
  quality_of_price: number;
  quality_of_friendliness: number;
  quality_of_timeschedule: number;
  provider_email: string;
  provider_id: number;
  company_name: string;
  project_cost: Array<any>;
  photo_urls: Array<any>;
  feedback_request_id: number;
}